backend/requirements/requirements-cpu.txt
fastapi==0.116.1
uvicorn==0.35.0
transformers==4.53.3
torch==2.7.1
Pillow==11.3.0
protobuf==6.31.1
sentencepiece==0.2.0
python-multipart==0.0.20
paddlepaddle==3.1.0
paddleocr==3.1.0
pydantic==2.11.7
redis==6.2.0
accelerate==1.9.0
asyncio==3.4.3
